Split Prod/Root Controller into provider/controller
